fix: split time and datefucntion
