+++ src/js/index.js
@@ -3,5 +3,6 @@ import '../sass/styles.scss';
 import '../sass/media.scss';
 import './location.js';
 import './date.js';
+import './time.js';
 import './weather.js';
 import './forecast.js';
